New application icon, splash, and installer images. Sampulse and Luahost fileformat Freeze. For sampulse, when saving EINS, the wavedata is not saved in INSD (avoiding to save it twice, but maintaining a bit of backward compatibility). Envelope mode (millis or ticks) visually implemented. Corrections in sampler envelopes (didn't show the correct value of milliseconds) Envelope ADSR/freeform now part of the envelope data and saved to file. Volume in sampulse instrument and sample bank in dB. Volume in sampler using the -inf..+12dB scale. LuaHost fixes for resampler tool. (cannot use the sse optimizations) Tweaks in native parameter view and mixer view without mouse movement (prevents the problem of tweaking knobs and reaching the top or the bottom of the screen with the mouse). Sliders still allow mouse movement, obviously.

+++ psycle-helpers/src/psycle/helpers/riffwave.cpp
@@ -526,7 +526,7 @@ if(hd.ckID == FourCC("smpl"))
 		for(uint32_t io = 0 ; io < samples ; io+=amount)
 		{
 			//truncate 1024 to amount of chans.
-			amount = std::min(32768U/chans,samples-io);
+			amount = static_cast<uint32_t>(std::min(32768U/chans,samples-io));
 			ReadArray(samps, amount*chans);
 			in_type* psamps = samps;
 			for (int a=0; a < amount; a++) {
